Moved BasicSelect options into separate component
